refactor to Next Pages router for gh-pages deployment
