--- app/layout.tsx
@@ -1,24 +1,0 @@
-import '@fortawesome/fontawesome-svg-core/styles.css';
-import type { Metadata, Viewport } from 'next';
-import './globals.css';
-
-export const metadata: Metadata = {
-  title: "Alex's Portfolio",
-  description: 'A personal portfolio for software engineer Alex Cravalho.',
-  icons: {
-    icon: '/favicon.png',
-  },
-};
-
-export const viewport: Viewport = {
-  width: 'device-width',
-  initialScale: 1,
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
-}
