+++ components/Education/education.data.ts
