+++ components/Skills/skills.data.ts
