+++ components/Software/software.data.ts
