+++ package.json
@@ -8,7 +8,8 @@
     "start": "next start",
     "lint": "next lint",
     "format": "prettier --write .",
-    "deploy": "gh-pages -d out"
+    "postbuild": "touch out/.nojekyll",
+    "deploy": "npm run build && gh-pages -d out --dotfiles"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
